Cmake script changes + a small permissions oopsie
Moved the script to the build directory as this way the script will put
the cmake files on there and not litter src directories with those
files. (I get why it was there beforehand)

Also it I accidentally the permissions on every file on this repo to 751 by
accident. This has been fixed.